ДЗ 2: Добавлены конфигурации
