+++ homework-02/src/main/java/ru/otus/homework/service/InputOutputServiceImpl.java
@@ -1,22 +1,18 @@
 package ru.otus.homework.service;
 
-import org.springframework.stereotype.Service;
-
 import java.io.InputStream;
 import java.io.PrintStream;
 import java.util.Scanner;
 
-@Service
 public class InputOutputServiceImpl implements InputOutputService {
-    private static final InputStream IN_STREAM = System.in;
-    private static final PrintStream OUT_STREAM = System.out;
 
     private final PrintStream outStream;
     private final Scanner scanner;
 
-    public InputOutputServiceImpl() {
-        this.outStream = OUT_STREAM;
-        this.scanner = new Scanner(IN_STREAM);
+    public InputOutputServiceImpl(InputStream inputStream,
+                                  PrintStream outStream) {
+        this.outStream = outStream;
+        this.scanner = new Scanner(inputStream);
     }
 
     public void showMessage(String message) {
